fix: router_main: Properly delete old config.
This was caused by configure + load not actually doing what I want,
not sure I'm certain why, but commit / exit at the end wouldn't of
worked because that would of screwed the current ssh session.

match: "none" seems like it was built for this case and I just
found out about it, so yeah. Let's use that. Cause I'm a dumbass
and didn't realize it existed.